+++ packages/core/src/integrations/third-party-errors-filter.ts
@@ -2,8 +2,8 @@ import { defineIntegration } from '../integration';
 import { addMetadataToStackFrames, stripMetadataFromStackFrames } from '../metadata';
 import type { EventItem } from '../types-hoist/envelope';
 import type { Event } from '../types-hoist/event';
-import { forEachEnvelopeItem } from '../utils-hoist/envelope';
-import { getFramesFromEvent } from '../utils-hoist/stacktrace';
+import { forEachEnvelopeItem } from '../utils/envelope';
+import { getFramesFromEvent } from '../utils/stacktrace';
 
 interface Options {
   /**