+++ packages/core/src/logs/exports.ts
@@ -5,11 +5,11 @@ import { DEBUG_BUILD } from '../debug-build';
 import type { Scope, ScopeData } from '../scope';
 import type { Log, SerializedLog, SerializedLogAttributeValue } from '../types-hoist/log';
 import { mergeScopeData } from '../utils/applyScopeDataToEvent';
+import { isParameterizedString } from '../utils/is';
+import { logger } from '../utils/logger';
 import { _getSpanForScope } from '../utils/spanOnScope';
-import { isParameterizedString } from '../utils-hoist/is';
-import { logger } from '../utils-hoist/logger';
-import { timestampInSeconds } from '../utils-hoist/time';
-import { GLOBAL_OBJ } from '../utils-hoist/worldwide';
+import { timestampInSeconds } from '../utils/time';
+import { GLOBAL_OBJ } from '../utils/worldwide';
 import { SEVERITY_TEXT_TO_SEVERITY_NUMBER } from './constants';
 import { createLogEnvelope } from './envelope';
 
@@ -138,14 +138,12 @@ export function _INTERNAL_captureLog(
     ...beforeLog.attributes,
   };
 
-  const { user } = getMergedScopeData(currentScope);
-  // Only attach user to log attributes if sendDefaultPii is enabled
-  if (client.getOptions().sendDefaultPii) {
-    const { id, email, username } = user;
-    setLogAttribute(processedLogAttributes, 'user.id', id, false);
-    setLogAttribute(processedLogAttributes, 'user.email', email, false);
-    setLogAttribute(processedLogAttributes, 'user.name', username, false);
-  }
+  const {
+    user: { id, email, username },
+  } = getMergedScopeData(currentScope);
+  setLogAttribute(processedLogAttributes, 'user.id', id, false);
+  setLogAttribute(processedLogAttributes, 'user.email', email, false);
+  setLogAttribute(processedLogAttributes, 'user.name', username, false);
 
   setLogAttribute(processedLogAttributes, 'sentry.release', release);
   setLogAttribute(processedLogAttributes, 'sentry.environment', environment);